clean up intrinsics module
